13.k8s/02.fastapi-k8s/fastapi_code/requirements.txt
fastapi
pydantic
uvicorn
pymongo
python-json-logger==2.0.7
python-logstash==0.4.8